+++ driver-core/src/main/java/com/datastax/driver/core/DefaultPreparedStatement.java
@@ -175,4 +175,9 @@ public class DefaultPreparedStatement implements PreparedStatement{
     public PreparedId getPreparedId() {
         return preparedId;
     }
+
+    @Override
+    public String toString() {
+        return "PreparedStatement[" + query + "]";
+    }
 }